bug fixing from int to long

+++ crudService/src/main/java/com/cs544/group7/crudService/repository/AirportRepository.java
@@ -6,6 +6,6 @@ import org.springframework.stereotype.Repository;
 import com.cs544.group7.crudService.domain.Airport;
 
 @Repository
-public interface AirportRepository extends JpaRepository<Airport, Integer> {
+public interface AirportRepository extends JpaRepository<Airport, Long> {
 	public Airport findByCode(String airportCode);
 }
